+++ pytorch_lightning/plugins/training_type/horovod.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 from contextlib import ExitStack
-from typing import Any, List, Optional, Tuple, Union
+from typing import List, Optional, Tuple
 
 import torch
 import torch.nn as nn
@@ -20,12 +20,12 @@ from torch.optim import Optimizer
 from torch.optim.lr_scheduler import _LRScheduler
 
 from pytorch_lightning.core.optimizer import LightningOptimizer
+from pytorch_lightning.plugins.collective.collective_plugin import Collective
+from pytorch_lightning.plugins.collective.horovod_collective import HorovodCollective
 from pytorch_lightning.plugins.io.checkpoint_plugin import CheckpointIO
 from pytorch_lightning.plugins.training_type.parallel import ParallelPlugin
 from pytorch_lightning.utilities import _HOROVOD_AVAILABLE
-from pytorch_lightning.utilities.distributed import distributed_available
-from pytorch_lightning.utilities.distributed import group as dist_group
-from pytorch_lightning.utilities.distributed import rank_zero_only, ReduceOp
+from pytorch_lightning.utilities.distributed import rank_zero_only
 
 if _HOROVOD_AVAILABLE:
     import horovod.torch as hvd
@@ -38,8 +38,14 @@ class HorovodPlugin(ParallelPlugin):
         self,
         parallel_devices: Optional[List[torch.device]] = None,
         checkpoint_io: Optional[CheckpointIO] = None,
+        collective: Optional[Collective] = None,
     ):
-        super().__init__(parallel_devices=parallel_devices, cluster_environment=None, checkpoint_io=checkpoint_io)
+        super().__init__(
+            parallel_devices=parallel_devices,
+            cluster_environment=None,
+            checkpoint_io=checkpoint_io,
+            collective=collective or HorovodCollective(),
+        )
         rank_zero_only.rank = self.global_rank
 
     @property
@@ -65,6 +71,8 @@ class HorovodPlugin(ParallelPlugin):
 
     def setup(self) -> None:
         self.model_to_device()
+        self.collective.on_gpu = self.on_gpu
+        self.collective.local_rank = self.local_rank
 
     def pre_dispatch(self):
 
@@ -108,14 +116,14 @@ class HorovodPlugin(ParallelPlugin):
             self._results = trainer.run_stage()
 
         # Make sure all workers have finished training before returning to the user
-        self.join()
+        self.collective.join()
 
     def start_evaluating(self, trainer):
         with ExitStack():
             self._results = trainer.run_stage()
 
         # Make sure all workers have finished training before returning to the user
-        self.join()
+        self.collective.join()
 
     def start_predicting(self, trainer):
         with ExitStack():
@@ -123,69 +131,13 @@ class HorovodPlugin(ParallelPlugin):
             self._results = trainer.run_stage()
 
         # Make sure all workers have finished training before returning to the user
-        self.join()
-
-    def barrier(self, *args, **kwargs):
-        if distributed_available():
-            self.join()
-
-    def broadcast(self, obj: object, src: int = 0) -> object:
-        obj = hvd.broadcast_object(obj, src)
-        return obj
+        self.collective.join()
 
     def model_to_device(self):
         if self.on_gpu:
             # this can potentially be removed after #8312. Not done due to lack of horovod testing
             torch.cuda.set_device(self.root_device)
         self.model.to(self.root_device)
-
-    def join(self):
-        if self.on_gpu:
-            hvd.join(self.local_rank)
-        else:
-            hvd.join()
-
-    def reduce(self, tensor, group: Optional[Any] = None, reduce_op: Optional[Union[ReduceOp, str]] = "mean"):
-        """Reduces a tensor from several distributed processes to one aggregated tensor.
-
-        Args:
-            tensor: the tensor to sync and reduce
-            group: the process group to gather results from. Defaults to all processes (world)
-            reduce_op: the reduction operation. Defaults to 'mean'/'avg'.
-                Can also be a string 'sum' to calculate the sum during reduction.
-
-        Return:
-            reduced value, except when the input was not a tensor the output remains is unchanged
-        """
-        if group is not None:
-            raise ValueError("Horovod does not support allreduce using a subcommunicator at this time. Unset `group`.")
-
-        if reduce_op in (None, "avg", "mean"):
-            reduce_op = hvd.Average
-        elif reduce_op in ("sum", ReduceOp.SUM):
-            reduce_op = hvd.Sum
-        else:
-            raise ValueError(f"unrecognized `reduce_op`: {reduce_op}")
-
-        # sync all processes before reduction
-        self.join()
-        return hvd.allreduce(tensor, op=reduce_op)
-
-    def all_gather(
-        self, result: Union[torch.Tensor], group: Optional[Any] = dist_group.WORLD, sync_grads: bool = False
-    ) -> torch.Tensor:
-        if group is not None and group != dist_group.WORLD:
-            raise ValueError("Horovod does not support allgather using a subcommunicator at this time. Unset `group`.")
-
-        if len(result.shape) == 0:
-            # Convert scalars to single dimension tensors
-            result = result.reshape(1)
-
-        # sync and gather all
-        self.join()
-        gathered = hvd.allgather(result)
-        gathered_result = list(gathered.split(1, dim=0))
-        return gathered_result
 
     def post_backward(self, closure_loss: torch.Tensor) -> None:
         # synchronize all horovod optimizers.