+++ pytorch_lightning/plugins/training_type/single_device.py
@@ -11,10 +11,12 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from typing import Any, Optional, Union
+from typing import Optional
 
 import torch
 
+from pytorch_lightning.plugins.collective.collective_plugin import Collective
+from pytorch_lightning.plugins.collective.single_device_collective import SingleDeviceCollective
 from pytorch_lightning.plugins.io.checkpoint_plugin import CheckpointIO
 from pytorch_lightning.plugins.training_type.training_type_plugin import TrainingTypePlugin
 from pytorch_lightning.utilities import _XLA_AVAILABLE
@@ -27,8 +29,9 @@ class SingleDevicePlugin(TrainingTypePlugin):
         self,
         device: torch.device,
         checkpoint_io: Optional[CheckpointIO] = None,
+        collective: Optional[Collective] = None,
     ):
-        super().__init__(checkpoint_io)
+        super().__init__(checkpoint_io=checkpoint_io, collective=collective or SingleDeviceCollective())
         self.device: torch.device = device
         self.global_rank = 0
         self.local_rank = 0
@@ -41,24 +44,6 @@ class SingleDevicePlugin(TrainingTypePlugin):
     @property
     def on_gpu(self) -> bool:
         return self.root_device.type == "cuda" and torch.cuda.is_available()
-
-    def reduce(self, tensor: Union[Any, torch.Tensor], *args: Any, **kwargs: Any) -> Union[Any, torch.Tensor]:
-        """Reduces a tensor from several distributed processes to one aggregated tensor. As this plugin only
-        operates with a single device, the reduction is simply the identity.
-
-        Args:
-            tensor: the tensor to sync and reduce
-            *args: ignored
-            **kwargs: ignored
-
-        Return:
-            the unmodified input as reduction is not needed for single process operation
-        """
-        return tensor
-
-    def all_gather(self, tensor: torch.Tensor, group: Optional[Any] = None, sync_grads: bool = False) -> torch.Tensor:
-        """Perform a all_gather on all processes."""
-        return tensor
 
     @property
     def root_device(self) -> torch.device:
@@ -73,12 +58,6 @@ class SingleDevicePlugin(TrainingTypePlugin):
     @property
     def is_global_zero(self) -> bool:
         return True
-
-    def barrier(self, *args, **kwargs) -> None:
-        pass
-
-    def broadcast(self, obj: object, src: int = 0) -> object:
-        return obj
 
     def teardown(self) -> None:
         if self.on_gpu: