+++ tests/core/test_metric_result_integration.py
@@ -316,7 +316,7 @@ def test_lightning_module_logging_result_collection(tmpdir, device):
             assert state_dict["items"]["validation_step.v"]["value"].device.type == device
 
             # sync fn should be kept
-            assert results["validation_step.v"].meta.sync.fn == self.trainer.training_type_plugin.reduce
+            assert results["validation_step.v"].meta.sync.fn == self.trainer.training_type_plugin.collective.reduce
 
             # sync fn dropped from the state dict
             assert "fn" not in state_dict["items"]["validation_step.v"]["meta"]["_sync"]
@@ -326,7 +326,7 @@ def test_lightning_module_logging_result_collection(tmpdir, device):
             assert results["validation_step.v"].value.device.type == device
 
             # sync fn was preserved in the original result
-            assert results["validation_step.v"].meta.sync.fn == self.trainer.training_type_plugin.reduce
+            assert results["validation_step.v"].meta.sync.fn == self.trainer.training_type_plugin.collective.reduce
 
             # default sync fn
             new_results = ResultCollection(False, device)
@@ -453,7 +453,7 @@ def result_collection_reload(**kwargs):
     assert not model.has_validated_sum
 
     tmpdir = (
-        trainer.training_type_plugin.broadcast(trainer_kwargs["default_root_dir"], 0)
+        trainer.training_type_plugin.collective.broadcast(trainer_kwargs["default_root_dir"], 0)
         if num_processes >= 2
         else trainer_kwargs["default_root_dir"]
     )