+++ tests/models/data/horovod/train_default_model.py
@@ -58,7 +58,9 @@ def run_test_from_config(trainer_options, on_gpu, check_size=True):
             assert self.device == expected_device
 
         def training_epoch_end(self, outputs) -> None:
-            res = self.trainer.training_type_plugin.reduce(torch.tensor(1.0, device=self.device), reduce_op="sum")
+            res = self.trainer.training_type_plugin.collective.reduce(
+                torch.tensor(1.0, device=self.device), reduce_op="sum"
+            )
             assert res.sum() == self.trainer.training_type_plugin.world_size
 
     model = TestModel()